Address review committee feedback (round 1)
simplicity-engineer requested deduping the isUnderForcedColors helper
between the Stylelint plugin and the parser-based recipe test. Extracted
to packages/components/scripts/stylelint/focus-ring-helpers.mjs with a
.d.mts declaration sidecar so the .ts test can import it through TS's
module resolver.

Also renamed the SideNavigationItemProps helper from WithoutVariant to
DistributiveOmit and added a docstring explaining why a plain
Omit<NavigationItemProps, 'variant'> would not work — Omit collapses the
link/button discriminated union and silently lets through extra props.

The other simplicity-engineer recommendations (collapse the Stylelint
rule's comma-branch parsing, drop the outline-none owner-comment escape
hatch, inline the type-test .svelte fixture) were rejected because the
plan explicitly mandates each of them and re-tests confirm the
implementation matches the plan's acceptance criteria.

Task: f73420f8-9c27-49e7-a245-37d7e2a339c3
